+++ src/lib.rs
@@ -453,7 +453,7 @@ mod script;
 mod writer;
 
 pub(crate) trait ExtraCheck<T: Sized = Self> {
-    fn check(_: &[(Option<T>, u32)]) -> Option<T> {
+    fn check(_: &[(Option<T>, usize)]) -> Option<T> {
         None
     }
 }
